requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
PyPDF2
python-dotenv
faiss-cpu
google-generativeai